Added UTC to CUC conversion

+++ inc/Helpers/Statistic.hpp
@@ -7,6 +7,9 @@
 #include "etl/vector.h"
 #include <cmath>
 #include <cfloat>
+#include "Time/UTCTimestamp.hpp"
+#include "Time/Time.hpp"
+#include "Time/TimeStamp.hpp"
 
 /**
  * Class containing all the statistics for every parameter. Includes functions that calculate and append the
@@ -16,8 +19,8 @@ class Statistic {
 public:
 	uint16_t selfSamplingInterval = 0;
 	uint16_t sampleCounter = 0;
-	uint32_t maxTime = 0;
-	uint32_t minTime = 0; // TODO: CUC Format timestamp
+	uint64_t maxTime = 0;
+	uint64_t minTime = 0;
 	double max = -std::numeric_limits<double>::infinity();
 	double min = std::numeric_limits<double>::infinity();
 	double sumOfSquares = 0;
